Add /chairs sit This command will allow you to sit on the block you are looking at, regardless of whether there are signs on the end
