+++ .gitignore
@@ -29,6 +29,11 @@ local.properties
 # PDT-specific
 .buildpath
 
+#################
+## Intellij IDEA
+#################
+.idea/
+Chairs.iml
 
 #################
 ## Visual Studio
