+++ src/com/cnaude/chairs/core/Utils.java
@@ -1,0 +1,271 @@
+package com.cnaude.chairs.core;
+
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
+import org.bukkit.entity.Entity;
+import org.bukkit.entity.ItemFrame;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.block.Action;
+import org.bukkit.event.player.PlayerInteractEvent;
+import org.bukkit.inventory.EquipmentSlot;
+import org.bukkit.material.Stairs;
+import org.bukkit.material.Step;
+import org.bukkit.material.WoodenStep;
+
+import com.cnaude.chairs.core.Chairs;
+
+public class Utils {
+
+    private final Chairs plugin;
+
+    public Utils(Chairs plugin) {
+        this.plugin = plugin;
+    }
+
+    public boolean sitAllowed(Player player, Block block, boolean signCheck) {
+
+        // Check for sitting disabled
+        if (plugin.sitDisabled.contains(player.getUniqueId())) {
+            return false;
+        }
+        if (plugin.disabledWorlds.contains(player.getWorld().getName())) {
+            return false;
+        }
+
+        // Check for permissions
+        if (!player.hasPermission("chairs.sit")) {
+            return false;
+        }
+
+        // Check for already sitting
+        if (isSitting(player)) {
+            return false;
+        }
+
+        // Check for item in hand
+        if (plugin.ignoreIfBlockInHand && player.getInventory().getItemInMainHand().getType() != Material.AIR) {
+            return false;
+        }
+
+        // Check for sneaking
+        if (player.isSneaking()) {
+            return false;
+        }
+
+        // Sit occupied check
+        if (plugin.getPlayerSitData().isBlockOccupied(block)) {
+            player.sendMessage(plugin.msgOccupied.replace("%PLAYER%", plugin.getPlayerSitData().getPlayerOnChair(block).getName()));
+            return false;
+        }
+
+        Stairs stairs = null;
+        Step step = null;
+        WoodenStep wStep = null;
+
+        // Check for block is chair
+        if (isValidChair(block)) {
+
+            if (block.getState().getData() instanceof Stairs) {
+                stairs = (Stairs) block.getState().getData();
+            } else if (block.getState().getData() instanceof Step) {
+                step = (Step) block.getState().getData();
+            } else if (block.getState().getData() instanceof WoodenStep) {
+                wStep = (WoodenStep) block.getState().getData();
+            }
+
+            int chairwidth = 1;
+
+            // Check if block beneath chair is solid.
+            if (block.getRelative(BlockFace.DOWN).isLiquid()) {
+                return false;
+            }
+            if (block.getRelative(BlockFace.DOWN).isEmpty()) {
+                return false;
+            }
+
+            // Check for distance distance between player and chair.
+            if (plugin.distance > 0 && player.getLocation().distance(block.getLocation().add(0.5, 0, 0.5)) > plugin.distance) {
+                return false;
+            }
+
+            // Check if block is inverted
+            if (stairs != null && stairs.isInverted()) {
+                return false;
+            }
+            if (step != null && step.isInverted()) {
+                return false;
+            }
+            if (wStep != null && wStep.isInverted()) {
+                return false;
+            }
+
+            // Check for signs (only for stairs)
+            if (signCheck && stairs != null) {
+                boolean sign1 = false;
+                boolean sign2 = false;
+
+                if (stairs.getDescendingDirection() == BlockFace.NORTH || stairs.getDescendingDirection() == BlockFace.SOUTH) {
+                    sign1 = checkSign(block, BlockFace.EAST) || checkFrame(block, BlockFace.EAST, player);
+                    sign2 = checkSign(block, BlockFace.WEST) || checkFrame(block, BlockFace.WEST, player);
+                } else if (stairs.getDescendingDirection() == BlockFace.EAST || stairs.getDescendingDirection() == BlockFace.WEST) {
+                    sign1 = checkSign(block, BlockFace.NORTH) || checkFrame(block, BlockFace.NORTH, player);
+                    sign2 = checkSign(block, BlockFace.SOUTH) || checkFrame(block, BlockFace.SOUTH, player);
+                }
+
+                if (!(sign1 && sign2)) {
+                    return false;
+                }
+            }
+
+            // Check for maximal chair width (only for stairs)
+            if (plugin.maxChairWidth > 0 && stairs != null) {
+                if (stairs.getDescendingDirection() == BlockFace.NORTH || stairs.getDescendingDirection() == BlockFace.SOUTH) {
+                    chairwidth += getChairWidth(block, BlockFace.EAST);
+                    chairwidth += getChairWidth(block, BlockFace.WEST);
+                } else if (stairs.getDescendingDirection() == BlockFace.EAST || stairs.getDescendingDirection() == BlockFace.WEST) {
+                    chairwidth += getChairWidth(block, BlockFace.NORTH);
+                    chairwidth += getChairWidth(block, BlockFace.SOUTH);
+                }
+
+                if (chairwidth > plugin.maxChairWidth) {
+                    return false;
+                }
+            }
+
+            return true;
+        }
+
+        return false;
+    }
+
+    public Location getSitLocation(Block block, Float playerYaw) {
+
+        Double sh = plugin.validChairs.get(block.getType());
+        if (sh == null) {
+            sh = 0.7;
+        }
+
+        Stairs stairs = null;
+        if (block.getState().getData() instanceof Stairs) {
+            stairs = (Stairs) block.getState().getData();
+        }
+
+        Location plocation = block.getLocation();
+        plocation.add(0.5D, (sh - 0.5D), 0.5D);
+
+        // Rotate the player's view to the descending side of the block.
+        if (plugin.autoRotate && stairs != null) {
+            switch (stairs.getDescendingDirection()) {
+                case NORTH: {
+                    plocation.setYaw(180);
+                    break;
+                }
+                case EAST: {
+                    plocation.setYaw(-90);
+                    break;
+                }
+                case SOUTH: {
+                    plocation.setYaw(0);
+                    break;
+                }
+                case WEST: {
+                    plocation.setYaw(90);
+                    break;
+                }
+                default: {
+                }
+            }
+        } else {
+            plocation.setYaw(playerYaw);
+        }
+        return plocation;
+    }
+
+
+
+    private boolean isValidChair(Block block) {
+        return plugin.validChairs.containsKey(block.getType());
+    }
+
+    private boolean isSitting(Player player) {
+        return plugin.getPlayerSitData().isSitting(player);
+    }
+
+    private int getChairWidth(Block block, BlockFace face) {
+        int width = 0;
+
+        // Go through the blocks next to the clicked block and check if there are any further stairs.
+        for (int i = 1; i <= plugin.maxChairWidth; i++) {
+            Block relative = block.getRelative(face, i);
+            if (relative.getState().getData() instanceof Stairs) {
+                if (isValidChair(relative) && ((Stairs) relative.getState().getData()).getDescendingDirection() == ((Stairs) block.getState().getData()).getDescendingDirection()) {
+                    width++;
+                } else {
+                    break;
+                }
+            }
+        }
+
+        return width;
+    }
+
+    private boolean checkSign(Block block, BlockFace face) {
+        // Go through the blocks next to the clicked block and check if are signs on the end.
+        for (int i = 1; i <= plugin.maxChairWidth + 2; i++) {
+            Block relative = block.getRelative(face, i);
+            if (checkDirection(block, relative)) {
+                continue;
+            }
+            if (plugin.validSigns.contains(relative.getType())) {
+                return true;
+            } else {
+                return false;
+            }
+        }
+        return false;
+    }
+
+    private boolean checkDirection(Block block1, Block block2) {
+        if (block1.getState().getData() instanceof Stairs && block2.getState().getData() instanceof Stairs) {
+            if (((Stairs) block1.getState().getData()).getDescendingDirection().equals(((Stairs) block2.getState().getData()).getDescendingDirection())) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private boolean checkFrame(Block block, BlockFace face, Player player) {
+        // Go through the blocks next to the clicked block and check if are signs on the end.
+
+        for (int i = 1; i <= plugin.maxChairWidth + 2; i++) {
+            Block relative = block.getRelative(face, i);
+            if (checkDirection(block, relative)) {
+                continue;
+            }
+            if (relative.getType().equals(Material.AIR)) {
+                int x = relative.getLocation().getBlockX();
+                int y = relative.getLocation().getBlockY();
+                int z = relative.getLocation().getBlockZ();
+                for (Entity e : player.getNearbyEntities(plugin.distance, plugin.distance, plugin.distance)) {
+                    if (e instanceof ItemFrame && plugin.validSigns.contains(Material.ITEM_FRAME)) {
+                        int x2 = e.getLocation().getBlockX();
+                        int y2 = e.getLocation().getBlockY();
+                        int z2 = e.getLocation().getBlockZ();
+                        if (x == x2 && y == y2 && z == z2) {
+                            return true;
+                        }
+                    }
+                }
+                return false;
+            } else {
+                return false;
+            }
+        }
+        return false;
+    }
+}
